requirements.txt
absl-py==1.0.0
asttokens==2.0.5
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
blis==0.7.5
cachetools==5.0.0
catalogue==2.0.6
catboost==1.0.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.1.0
colorama==0.4.4
cycler==0.11.0
cymem==2.0.6
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
executing==0.8.2
filelock==3.6.0
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.19.1
grpcio==1.44.0
h5py==3.6.0
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.1
iniconfig==1.1.1
ipython==8.1.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
keras-tuner==1.1.0
kiwisolver==1.3.2
kt-legacy==1.0.4
langcodes==3.3.0
libclang==13.0.0
llvmlite==0.38.1
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
murmurhash==1.0.6
numba==0.55.2
numpy==1.22.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
parso==0.8.3
pathy==0.6.1
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.6.0
pluggy==1.0.0
preshed==3.0.6
prompt-toolkit==3.0.28
protobuf==3.19.4
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
Pygments==2.11.2
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.8.0
sentencepiece==0.1.96
six==1.16.0
slicer==0.0.7
smart-open==5.2.1
spacy==3.2.2
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
stack-data==0.2.0
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.0.13
threadpoolctl==3.1.0
tokenizers==0.11.5
tomli==2.0.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.16.2
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.8
wasabi==0.9.0
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
